+++ Buildenator/Configuration/DefaultConstants.cs
@@ -8,4 +8,5 @@ internal static class DefaultConstants
     public const string ValueLiteral = "value";
     public const string NullBox = "NullBox";
     public const string FixtureLiteral = "_fixture";
+    public const string BuildMethodName = "Build";
 }